исправлены замечания преподавателя, код разнесен по модулям
